Always display most recent messages
By the way blessed-contrib's log widget works, messages added to the
widget first may not be displayed in the list of messages, if the screen
is not able to fit the amount of lines intended.

As we sort messages in descending order by default, the current
implementation may hide some of the most recent messages for the
stream. To fix it, we sort the messages by timestamp before adding them.
That will change the default order of the list of messages.